add a component for dialog
